requirements.txt
Pillow==9.0.0
scipy==1.7.3
streamlit==1.5.0
click==8.0